cleaned: models

+++ core/model/src/main/java/eu/golovkov/core/model/data/Player.kt
@@ -2,11 +2,11 @@ package eu.golovkov.core.model.data
 
 data class Player(
     val firstName: String,
-    val heightFeet: Int,
-    val heightInches: Int,
+    val heightFeet: Float,
+    val heightInches: Float,
     val id: Int,
     val lastName: String,
     val position: String,
     val team: Team,
-    val weightPounds: Int
+    val weightPounds: Float
 )
